requirements.txt
absl-py==2.3.1
adispatch==0.5.0
aiobotocore==2.19.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.16
altair==5.5.0
anaconda-anon-usage==0.7.1
anaconda-auth==0.8.6
anaconda-catalogs==0.2.0
anaconda-cli-base==0.5.2
anaconda-client==1.13.0
anaconda-navigator==2.6.6
anaconda-project==0.11.1
annotated-types==0.6.0
anyio==4.7.0
appdirs==1.4.4
archspec==0.2.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.11.0
astroid==3.3.8
astropy==7.0.0
astropy-iers-data==0.2025.1.13.0.34.51
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.17.0
atomicwrites==1.4.0
attrs==24.3.0
autobahn==25.12.2
Automat==24.8.1
autopep8==2.0.4
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.10.0
bleach==6.2.0
blinker==1.9.0
bokeh==3.6.2
boltons==24.1.0
botocore==1.36.3
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.1
cbor2==5.8.0
certifi==2025.10.5
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.1
conda==25.5.1
conda-anaconda-telemetry==0.1.2
conda-anaconda-tos==0.2.0
conda-build==25.5.0
conda-content-trust==0.2.0
conda_index==0.6.1
conda-libmamba-solver==25.4.0
conda-pack==0.7.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
conda-repo-cli==1.0.165
conda-token==0.6.0
constantly==23.10.4
contourpy==1.3.1
cookiecutter==1.7.3
cryptography==44.0.1
cssselect==1.2.0
cycler==0.11.0
cytoolz==1.0.1
daphne==4.2.1
dask==2025.2.0
dask-expr==2.0.0
datashader==0.18.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.8
distributed==2025.2.0
distro==1.9.0
Django==6.0
docker==7.1.0
docstring-to-markdown==0.11
docutils==0.21.2
et_xmlfile==1.1.0
evalidate==2.0.3
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.17.0
flake8==7.1.1
Flask==3.1.0
flatbuffers==25.9.23
fonttools==4.55.3
frozendict==2.4.2
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.7
GitPython==3.1.43
gmpy2==2.2.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.75.1
h11==0.16.0
h5py==3.12.1
HeapDict==1.0.1
holoviews==1.20.2
httpcore==1.0.9
httpx==0.28.1
hvplot==0.11.3
hyperlink==21.0.0
idna==3.7
imageio==2.37.0
imagesize==1.4.1
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
incremental==24.7.2
inflection==0.5.1
iniconfig==1.1.1
intake==2.0.7
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isort==6.0.1
itemadapter==0.3.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.2.1
jaraco.context==0.0.0
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.8.2
kagglesdk==0.1.13
keras==3.11.3
keyring==25.6.0
kiwisolver==1.4.8
lazy_loader==0.4
lckr_jupyterlab_variableinspector==3.2.4
libarchive-c==5.1
libclang==18.1.1
libmambapy==2.0.5
lief==0.16.4
linkify-it-py==2.0.0
llvmlite==0.44.0
lmdb==1.6.2
locket==1.0.0
lxml==5.3.0
lz4==4.3.2
Markdown==3.8
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.0
menuinst==2.2.0
mistune==3.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml_dtypes==0.5.3
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multipledispatch==0.6.0
mypy==1.18.2
mypy_extensions==1.0.0
namex==0.1.0
narwhals==1.31.0
navigator-updater==0.5.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numexpr==2.10.1
numpy==2.1.3
numpydoc==1.7.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
panel==1.7.0
param==2.2.0
parsel==1.8.1
parso==0.8.4
partd==1.4.2
pathspec==0.10.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.1
pkce==1.0.3
pkginfo==1.12.0
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.5.0
ply==3.11
poyo==0.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.43
propcache==0.3.1
Protego==0.4.0
protobuf==5.29.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-ubjson==0.16.1
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.12.1
pycosat==0.6.6
pycparser==2.21
pyct==0.5.0
pycurl==7.45.6
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.6.1
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyerfa==2.0.1.5
pyflakes==3.2.0
PyGithub==2.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==5.0.2
pytoolconfig==1.2.6
pytz==2024.1
pyuca==1.2
pyviz_comms==3.0.2
PyWavelets==1.8.0
pywin32==308
pywin32-ctypes==0.2.2
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.4.0
qtconsole==5.6.1
QtPy==2.4.1
queuelib==1.6.2
readchar==4.0.5
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rich==13.9.4
roman-numerals-py==3.1.0
rope==1.13.0
rpds-py==0.22.3
Rtree==1.0.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruamel_yaml_conda==0.17.21
s3fs==2025.3.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.12.0
seaborn==0.13.2
semver==3.0.2
Send2Trash==1.8.2
service-identity==24.2.0
setuptools==72.1.0
shellingham==1.5.0
sip==6.7.12
six==1.17.0
sklearn-compat==0.1.3
smmap==4.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.7
spyder-kernels==3.0.5
SQLAlchemy==2.0.39
sqlparse==0.5.5
stack-data==0.2.0
statsmodels==0.14.4
streamlit==1.45.1
superqt==0.7.3
sympy==1.13.3
tables==3.10.2
tabulate==0.9.0
tblib==3.1.0
tenacity==9.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.17.1
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.5.0
three-merge==0.1.1
tifffile==2025.2.18
tinycss2==1.4.0
tldextract==5.1.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.0
Twisted==24.11.0
twisted-iocpsupport==1.0.2
txaio==25.12.2
typer==0.9.0
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.1
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.3.0
uv==0.9.7
uvicorn==0.40.0
w3lib==2.1.2
watchdog==4.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whatthepatch==1.0.2
wheel==0.45.1
widgetsnbextension==4.0.13
win_inet_pton==1.1.0
wrapt==1.17.0
xarray==2025.4.0
xgboost==3.0.5
xlwings==0.32.1
xyzservices==2022.9.0
yapf==0.40.2
yarl==1.18.0
zict==3.0.0
zipp==3.21.0
zope.interface==7.1.1
zstandard==0.23.0
